requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-bigquery==2.6.1
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
urllib3==1.26.2
